refactor(privacy): move scan reporting beside the programs and capture the warehouse intro menu
The consent-gated reporting helper is program knowledge, so it lives next to
the other cross-program gates under src/lib/programs rather than inside the
pure detection directory. The warehouse intro screen now captures its menu
selection like the install screen does, so the decline rate is measurable
for both programs.
